pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cell-os"
version = "0.1.0"
description = "Autonomous lab operating system for cell biology experiments"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "scipy",
    "pyyaml",
    "streamlit",
    "matplotlib",
    "graphviz",
    "altair",
    "plotly",
    "networkx",
    "ortools",
    "tqdm",
    "fastapi",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff", "mypy"]

[project.scripts]
cell-os-run = "cell_os.cli.run_campaign:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselected by default, use -m slow to run)",
    "manual: marks tests for manual execution only (skipped in CI, use -m manual to run)",
    "realism: statistical property checks (not default CI, use -m realism to run)",
]
addopts = "-m 'not manual'"
